+++ src/components/index.js
@@ -1,14 +1,9 @@
 import DeclareSetup from './DeclareSetupComponent';
-import S3AnalyticsTable from './S3AnalyticsTableComponent';
-import S3AnalyticsInfos from './S3AnalyticsInfosComponent';
-import S3AnalyticsBarChart from './S3AnalyticsBarChartComponent';
-
+import S3Analytics from './s3Analytics';
 import Misc from './misc';
 
 export default {
   DeclareSetup,
   Misc,
-  S3AnalyticsTable,
-  S3AnalyticsInfos,
-  S3AnalyticsBarChart,
+  S3Analytics,
 };
