+++ src/components/s3Analytics/S3AnalyticsBarChartComponent.js
