+++ src/components/s3Analytics/S3AnalyticsTableComponent.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import ReactTable from "react-table";
 
-import { formatBytes, formatPrice } from '../common/formatters';
+import { formatBytes, formatPrice } from '../../common/formatters';
 
 
 // S3AnalyticsTableComponent Component
