+++ src/components/s3Analytics/index.js
@@ -1,0 +1,9 @@
+import S3AnalyticsBarChart from './S3AnalyticsBarChartComponent';
+import S3AnalyticsInfos from './S3AnalyticsInfosComponent';
+import S3AnalyticsTable from './S3AnalyticsTableComponent';
+
+export default {
+  S3AnalyticsBarChart,
+  S3AnalyticsTable,
+  S3AnalyticsInfos,
+};
