+++ src/containers/s3AnalyticsContainer.js
@@ -29,19 +29,19 @@ class S3AnalyticsContainer extends Component {
         <div className="row">
           <div className="col-md-12">
             <div className="white-box">
-              {this.props.s3Data && <Components.S3AnalyticsInfos data={this.props.s3Data}/>}
+              {this.props.s3Data && <Components.S3Analytics.S3AnalyticsInfos data={this.props.s3Data}/>}
             </div>
           </div>
           <div className="col-md-12">
             <div className="white-box">
-              {this.props.s3Data && <Components.S3AnalyticsBarChart elementId="s3BarChart" data={this.props.s3Data}/>}
+              {this.props.s3Data && <Components.S3Analytics.S3AnalyticsBarChart elementId="s3BarChart" data={this.props.s3Data}/>}
             </div>
           </div>
 
         </div>
 
         <div className="white-box no-padding">
-          {this.props.s3Data && <Components.S3AnalyticsTable data={this.props.s3Data}/>}
+          {this.props.s3Data && <Components.S3Analytics.S3AnalyticsTable data={this.props.s3Data}/>}
         </div>
       </div>
     );
