+++ src/index.js
@@ -19,6 +19,7 @@ import './styles/style.css';
 import './styles/style-sm.css';
 import './styles/style-xs.css';
 import './styles/style-md.css';
+import './styles/helpers.css';
 
 // Components
 import App from './App';
