+++ src/styles/style.css
@@ -114,114 +114,7 @@ button {
 }
 
 
-
-/* MARGIN AND PADDING HELPERS */
-.no-padding {
-  padding: 0px !important;
-}
-
-.no-margin {
-  margin: 0px !important;
-}
-
-.m-b-0 {
-    margin-bottom: 0px;
-}
-
-.m-l-0 {
-    margin-left: 0px;
-}
-
-.m-t-0 {
-    margin-top: 0px;
-}
-
-.m-r-0 {
-    margin-right: 0px;
-}
-
-.p-b-0 {
-    padding-bottom: 0px;
-}
-
-.p-l-0 {
-    padding-left: 0px;
-}
-
-.p-t-0 {
-    padding-top: 0px;
-}
-
-.p-r-0 {
-    padding-right: 0px;
-}
-
-.m-b-10 {
-    margin-bottom: 10px;
-}
-
-.m-l-10 {
-    margin-left: 10px;
-}
-
-.m-t-10 {
-    margin-top: 10px;
-}
-
-.m-r-10 {
-    margin-right: 10px;
-}
-
-
-.p-b-10 {
-    padding-bottom: 10px;
-}
-
-.p-l-10 {
-    padding-left: 10px;
-}
-
-.p-t-10 {
-    padding-top: 10px;
-}
-
-.p-r-10 {
-    padding-right: 10px;
-}
-
-.m-b-15 {
-    margin-bottom: 15px;
-}
-
-.m-l-15 {
-    margin-left: 15px;
-}
-
-.m-t-15 {
-    margin-top: 15px;
-}
-
-.m-r-15 {
-    margin-right: 15px;
-}
-
-
-.p-b-15 {
-    padding-bottom: 15px;
-}
-
-.p-l-15 {
-    padding-left: 15px;
-}
-
-.p-t-15 {
-    padding-top: 15px;
-}
-
-.p-r-15 {
-    padding-right: 15px;
-}
-
+/* WHITE BOX */
 
 .white-box {
     background: #fff;
